DCEC model xbatch casting to float32

+++ src/models/DCEC/models/DCEC_model.py
@@ -167,7 +167,7 @@ class DCECModel(BaseModel):
 
         """
         tensor = input[0]
-        self.x_batch = tensor.to(self.device) # copy
+        self.x_batch = tensor.type(torch.FloatTensor).to(self.device) # copy
 
     def encode(self):
         """Run encoding pass"""
